// frontend/src/types/openai.d.ts
export interface components {
  schemas: {
// ...
    CreateChatCompletionStreamResponse: {
      id: string;
      object: string;
      created: number;
      model: string;
      choices: ({
          index?: number;
          delta?: components["schemas"]["ChatCompletionStreamResponseDelta"];
          /** @enum {string} */
          finish_reason?: "stop" | "length" | "function_call" | null;
        })[];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}